[FIX] Code smell

+++ src/solver/variable/state.h
@@ -60,7 +60,7 @@ public:
         //! Minimum power of all clusters for the current hour in the year
         std::vector<double> PMinOfClusters;
 
-        std::vector<double> unitCountLastHour;
+        std::vector<uint> unitCountLastHour;
         std::vector<double> productionLastHour;
         std::vector<double> pminOfAGroup;
     };
